Remove single-use MetadataObject class
